+++ interface/scripted/starcustomchat/base/starcustomchatgui.lua
@@ -319,6 +319,29 @@ function registerCallbacks()
     self.customChat:clearHistory()
   end))
 
+    starcustomchat.utils.setMessageHandler("scc_edit_message", function(_, _, data)
+    local msgInd = self.customChat:findMessageByUUID(data.uuid)
+    if msgInd then
+      data = self.customChat.callbackPlugins("formatIncomingMessage", data)
+      local message = self.customChat.messages[msgInd]
+      message.text = data.text
+      message.mode = data.mode
+      message.textHeight = nil
+      if not message.edited then
+        message.time = "^set;^lightgray;(" .. starcustomchat.utils.getTranslation("chat.message.edited") .. ")^reset; "
+          .. (message.time or "")
+        message.edited = true
+        message.forceAvatar = true
+      end
+
+      local newUUID = util.hashString(data.connection .. data.text)
+      local oldUUID = message.uuid
+      self.customChat:replaceUUID(oldUUID, newUUID)
+
+      self.customChat:processQueue()
+    end
+  end)
+
   self.runCallbackForPlugins("registerMessageHandlers")
 
   return true