+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Features
 + Rearranged some context menu buttons to prevent accidental pings etc.
 + Added new buttons for left menu and plugins in settings to indicate the scroll area (work in newer versions of OpenStarbound)
++ Moved the handler of editted messages to the base mod
 
 ## API
 + Introduced internal forceAvatar parameter